Update monitor for nimiq rpc 0.1.3

+++ monitor/src/types.rs
@@ -1,6 +1,6 @@
 use thiserror::Error;
 
-pub const ACTIVATION_HEIGHT: u64 = 100;
+pub const ACTIVATION_HEIGHT: u32 = 100;
 pub const BURN_ADDRESS: &str = "NQ07 0000 0000 0000 0000 0000 0000 0000 0000";
 pub enum ValidatorsReadiness {
     NotReady(u16),
